pdf-fe/pyproject.toml
[project]
name = "pdf-fe"
version = "0.1.0"
description = ""
authors = [
    {name = "lorenzobenzoni",email = "lorenzo1.benzoni@mail.polimi.it"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.42.0,<2.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
